+++ core/components/admin/courses/evaluation/edit/CodeCompletionEditor.tsx
